move FakeImage test class into a support file
Both ec2_spec and ec2/image_selection_spec use this class, so let's
promote it to a proper shared test class under spec/support.
